refactor: mover ventas a features/ventas/
- Mover componentes, domain types, hooks y páginas de ventas
  a src/features/ventas/ siguiendo estructura feature-based
- Actualizar imports en routes, layout, constants y CatalogoProductos
- Eliminar carpetas obsoletas src/components/ventas/ y src/pages/
