+++ locuventas_frontend/src/features/ventas/domain/venta.types.ts
