+++ locuventas_frontend/src/features/ventas/pages/Dashboard.tsx
@@ -1,17 +1,17 @@
 import { useState, useEffect } from "react";
-import type { Venta, VentaDetalle } from "@domain/venta.types";
-import type { CarritoItem } from "@hooks/useCarrito";
+import type { Venta, VentaDetalle } from "../domain/venta.types";
+import type { CarritoItem } from "../hooks/useCarrito";
 import AppLayout from "@layout/AppLayout";
 import Aside from "@layout/Aside";
 import Main from "@layout/Main";
 import { useHeader } from "@context/HeaderContext";
-import CarritoVenta from "@components/ventas/CarritoVentas";
+import CarritoVenta from "../components/CarritoVentas";
 import CatalogoProductos from "@features/productos/components/CatalogoProductos";
-import ModalPago from "@components/ventas/ModalPago";
-import ModalDetalleVenta from "@components/ventas/ModalDetalleVenta";
+import ModalPago from "../components/ModalPago";
+import ModalDetalleVenta from "../components/ModalDetalleVenta";
 import { apiRequest } from "@services/api";
-import { useCarrito } from "@hooks/useCarrito";
-import DrawerCarrito from "@components/ventas/DrawerCarrito";
+import { useCarrito } from "../hooks/useCarrito";
+import DrawerCarrito from "../components/DrawerCarrito";
 import useBreakpoint from "@hooks/useBreakpoint";
 import { isBreakpoint, BREAKPOINTS } from "@constants/breakpoints";
 import { toast } from "react-toastify";
